refactor: use timeslot hook.
